+++ src/test/java/it/gov/pagopa/standinmanager/AsyncServiceTest.java
@@ -1,13 +1,9 @@
 package it.gov.pagopa.standinmanager;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
-
 import com.azure.cosmos.CosmosClient;
 import com.azure.cosmos.CosmosContainer;
 import com.azure.cosmos.CosmosDatabase;
 import com.azure.cosmos.models.SqlQuerySpec;
-import com.azure.cosmos.util.CosmosPagedFlux;
 import com.azure.cosmos.util.CosmosPagedIterable;
 import com.microsoft.azure.kusto.data.Client;
 import com.microsoft.azure.kusto.data.exceptions.DataClientException;
@@ -26,23 +22,25 @@ import it.gov.pagopa.standinmanager.repository.model.CosmosStandInStation;
 import it.gov.pagopa.standinmanager.service.AsyncService;
 import it.gov.pagopa.standinmanager.service.ConfigService;
 import it.gov.pagopa.standinmanager.service.StationMonitorService;
-import java.time.Instant;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
 
+import java.time.Instant;
+import java.time.ZonedDateTime;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
+
 @ExtendWith(MockitoExtension.class)
-class StationMonitorServiceTest {
+class AsyncServiceTest {
 
     @Mock private Client kustoClient;
     @Mock private CosmosClient cosmosClient;
@@ -59,11 +57,8 @@ class StationMonitorServiceTest {
 
     private ForwarderClient forwarderClient = new ForwarderClient();
 
-    @Mock
-    AsyncService asyncService;
-
     @InjectMocks
-    private StationMonitorService stationMonitorService;
+    private AsyncService asyncService;
 
     @BeforeEach
     void setUp() throws KustoServiceQueryError, DataServiceException, DataClientException {
@@ -100,34 +95,37 @@ class StationMonitorServiceTest {
         stationCreditorInstitutions.put(station2.getStationCode() + "_creditorInstitution",stationCreditorInstitution1);
         configDataV1.setCreditorInstitutionStations(stationCreditorInstitutions);
 
-        when(configService.getCache()).thenReturn(configDataV1);
+//        when(configService.getCache()).thenReturn(configDataV1);
 //        when(restTemplate.exchange(any(), any(Class.class))).thenReturn(new ResponseEntity<String>("", HttpStatus.OK));
         when(cosmosClient.getDatabase(any())).thenReturn(cosmosDatabase);
         when(cosmosDatabase.getContainer(any())).thenReturn(cosmosContainer);
-        when(cosmosContainer.queryItems(any(SqlQuerySpec.class),any(),any())).thenReturn(cosmosPagedIterable);
-        when(cosmosPagedIterable.stream()).thenReturn(Arrays.asList(
-                new CosmosStandInStation("","station1",Instant.now()),
-                new CosmosStandInStation("","station2",Instant.now())
-        ).stream());
+//        when(cosmosContainer.queryItems(any(SqlQuerySpec.class),any(),any())).thenReturn(cosmosPagedIterable);
+//        when(cosmosPagedIterable.stream()).thenReturn(Arrays.asList(
+//                new CosmosStandInStation("","station1",Instant.now()),
+//                new CosmosStandInStation("","station2",Instant.now())
+//        ).stream());
 
         org.springframework.test.util.ReflectionTestUtils.setField(cosmosNodeDataRepository, "cosmosClient", cosmosClient);
-//        org.springframework.test.util.ReflectionTestUtils.setField(cosmosStationDataRepository, "cosmosClient", cosmosClient);
+        org.springframework.test.util.ReflectionTestUtils.setField(cosmosStationDataRepository, "cosmosClient", cosmosClient);
         org.springframework.test.util.ReflectionTestUtils.setField(cosmosStationRepository, "cosmosClient", cosmosClient);
         org.springframework.test.util.ReflectionTestUtils.setField(cosmosEventsRepository, "cosmosClient", cosmosClient);
-//        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "cosmosEventsRepository", cosmosEventsRepository);
-//        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "url", "http://forwarder.it");
-//        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "key", "key");
-//        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "restTemplate", restTemplate);
+        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "cosmosEventsRepository", cosmosEventsRepository);
+        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "url", "http://forwarder.it");
+        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "key", "key");
+        org.springframework.test.util.ReflectionTestUtils.setField(forwarderClient, "restTemplate", restTemplate);
 //        org.springframework.test.util.ReflectionTestUtils.setField(stationMonitorService, "forwarderClient", forwarderClient);
 //        org.springframework.test.util.ReflectionTestUtils.setField(stationMonitorService, "cosmosStationDataRepository", cosmosStationDataRepository);
-        org.springframework.test.util.ReflectionTestUtils.setField(stationMonitorService, "cosmosStationRepository", cosmosStationRepository);
+//        org.springframework.test.util.ReflectionTestUtils.setField(stationMonitorService, "cosmosStationRepository", cosmosStationRepository);
     }
 
     @Test
     void test1() throws Exception {
-        stationMonitorService.checkStations();
-        Thread.sleep(2000);
-//        verify(cosmosStationDataRepository, times(2)).save(any());
-        verify(asyncService, times(2)).checkStation(any(), any(), any(), any());
+        Station station  = new Station();
+        station.setStationCode("station1");
+
+        StationCreditorInstitution stationCreditorInstitution = new StationCreditorInstitution();
+
+        asyncService.checkStation(ZonedDateTime.now(), station, stationCreditorInstitution, CosmosStandInStation.builder().build());
+        verify(cosmosStationDataRepository, times( 1)).save(any());
     }
 }
